real offline first

+++ app/src/main/java/com/example/stockapp/ui/ReceitasViewModel.kt
@@ -1,21 +1,24 @@
 package com.example.stockapp.ui
 
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.Network
+import android.net.NetworkCapabilities
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.stockapp.data.daos.ReceitaDao
 import com.example.stockapp.data.objects.Receita
 import com.example.stockapp.data.repositories.ReceitaRepository
 import com.example.stockapp.data.repositories.ReceitaRepositorySQlite
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
 class ReceitasViewModel
-@Inject constructor(val repository: ReceitaRepository, val localRepository: ReceitaRepositorySQlite) : ViewModel() {
+@Inject constructor(val repository: ReceitaRepository, val localRepository: ReceitaRepositorySQlite, val context: Context, val receitasDao: ReceitaDao) : ViewModel() {
 
     var receita: Receita = Receita()
 
@@ -25,10 +28,29 @@ class ReceitasViewModel
     private var isUpdatingLocalData = false
 
     init {
+
+        print("INICIANDO RECEITA VIEW MODEL")
+        loadBase()
+        iniciarObservacaoConectividade()
+    }
+
+    private fun iniciarObservacaoConectividade() {
+        val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+
+        val networkCallback = object : ConnectivityManager.NetworkCallback() {
+            override fun onAvailable(network: Network) {
+                super.onAvailable(network)
+                syncWithFirebase()
+            }
+        }
+
+        connectivityManager.registerDefaultNetworkCallback(networkCallback)
+    }
+
+    fun loadBase() {
         viewModelScope.launch {
             repository.receitas.collect { firebaseReceitas ->
                 if (!isUpdatingLocalData) {
-                    println("FIREBASE INIT")
                     isUpdatingLocalData = true
                     localRepository.updateLocalData(firebaseReceitas)
                     isUpdatingLocalData = false
@@ -39,11 +61,38 @@ class ReceitasViewModel
         viewModelScope.launch {
             localRepository.receitas.collect { localReceitas ->
                 if (!isUpdatingLocalData) {
-                    println("LOCAL INIT")
                     _receitas.value = localReceitas
                 }
             }
         }
+    }
+
+    private fun syncWithFirebase() = viewModelScope.launch {
+        // Obtém todas as receitas não sincronizadas do banco local
+
+        val unsyncedReceitas = localRepository.getUnsyncedReceitas()
+
+        println("Pegando receitas não syncadas" + unsyncedReceitas)
+
+        // Sincroniza cada receita não sincronizada com o Firebase
+        for (receita in unsyncedReceitas) {
+            if (isNetworkAvailable(context)) {
+                receita.isSynced = true
+                repository.set(receita)
+
+                if (receita.isDeleted) {
+                    repository.delete(receita)
+                }
+
+                println("Sincronizado com o Firebase: $receita")
+            } else {
+                println("Sem conexão de internet. Sincronização adiada.")
+                return@launch
+            }
+        }
+
+        // Atualiza a flag isSynced para true no banco local
+        localRepository.updateSyncStatus(unsyncedReceitas, true)
     }
 
     fun new() {
@@ -58,16 +107,45 @@ class ReceitasViewModel
     fun set() = viewModelScope.launch {
         println("receita saved in database")
         println(receita)
+
+        receita.isSynced = false
         localRepository.set(receita)
-        repository.set(receita)
+
+        if (isNetworkAvailable(context)) {receita.isSynced = true
+            repository.set(receita)
+            println("passou pelo Firebase")
+        } else {
+            println("Sem conexão de internet. Operação no Firebase adiada.")
+        }
 
         new()
     }
 
     fun delete(receita: Receita) = viewModelScope.launch {
         println("receita deleted from database")
-        repository.delete(receita)
+
+        receita.isSynced = false
         localRepository.delete(receita)
+
+        if (isNetworkAvailable(context)) {
+            repository.delete(receita)
+            receita.isSynced = true
+            println("passou pelo Firebase")
+        } else {
+            println("Sem conexão de internet. Operação no Firebase adiada.")
+        }
+
         new()
+    }
+
+    private fun isNetworkAvailable(context: Context): Boolean {
+        val connectivityManager =
+            context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+        val network = connectivityManager.activeNetwork
+        val capabilities =
+            connectivityManager.getNetworkCapabilities(network)
+        return capabilities != null &&
+                (capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) ||
+                        capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR))
     }
 }
